requirements.txt
# Core
gradio 
pydantic 
requests 

# LangChain / LangGraph for structured outputs
langchain 
langchain-openai 
langchain-community 
langgraph 

# OpenEnv (hackathon framework)
openenv-core

# Optional: faster JSON
orjson 

# Existing deps
matplotlib
numpy 